+++ transports/websocket/CHANGELOG.md
@@ -1,6 +1,26 @@
-# 0.26.0 [unreleased]
+# 0.28.0 [2021-01-12]
+
+- Update dependencies.
+
+# 0.27.0 [2020-12-17]
 
 - Update `libp2p-core`.
+
+# 0.26.3 [2020-12-10]
+
+- Update `async-tls`.
+
+# 0.26.2 [2020-12-09]
+
+- Update minimum patch version for `async-tls`.
+
+# 0.26.1 [2020-12-07]
+
+- Update `rustls`.
+
+# 0.26.0 [2020-11-25]
+
+- Update dependencies.
 
 # 0.25.0 [2020-11-09]